Moved HawtioConfiguration to springboot.
Moved the Configuration class to the springboot module, and updated the sample project.
