+++ hawtio-sample-springboot/pom.xml
@@ -44,16 +44,6 @@
             <artifactId>hawtio-core</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>io.hawt</groupId>
-            <artifactId>hawtio-system</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.hawt</groupId>
-            <artifactId>hawtio-plugin-mbean</artifactId>
-            <version>${project.version}</version>
-        </dependency>
     </dependencies>
 
     <build>
@@ -75,8 +65,6 @@
                 <version>2.6</version>
                 <configuration>
                     <failOnMissingWebXml>false</failOnMissingWebXml>
-                    <dependentWarIncludes></dependentWarIncludes>
-                    <dependentWarExcludes>**</dependentWarExcludes>
                 </configuration>
             </plugin>
             <plugin>
